requirements.txt
python-arango==3.12.1
Django==1.10
psycopg2 >= 2.5.4
djangorestframework==3.5
djangorestframework-jwt==1.8.0
coverage==3.6
Faker
django-haystack>=2.5.dev1
elasticsearch==2.4
gunicorn
python-jose
drf-haystack
sendgrid
requests
django-cors-headers
django-split-settings
cerberus
ldap3
pysaml2==4.0.5
django_tools
boto3==1.4.5
django_elasticsearch_dsl
elasticsearch-dsl>=2.0,<3.0
botocore==1.7.10
docutils==0.14
futures==3.1.1
jmespath==0.9.3
numpy==1.13.1
pandas==0.20.3
python-dateutil==2.6.1
pytz==2017.2
s3transfer==0.1.11
six==1.10.0
SQLAlchemy==1.1.14
django-filter
pyspark
sqlalchemy-migrate==0.11.0
ndg-httpsclient
celery==4.1.0
raven==6.4.0
